zarza_etc::curve
Trait BinaryCurveOperations

pub trait BinaryCurveOperations<T> {
    // Required methods
    fn scale_axis(&mut self, axis: CurveAxis, by_what: T);
    fn multiply(&mut self, by_what: T);
    fn add(&mut self, what: T);
}

Trait defining binary operations that can be performed on a Curve: Operations that involve two curves or a curve and a scalar. These operations modify the curve in place.

Required Methods

fn scale_axis(&mut self, axis: CurveAxis, by_what: T)

Scales a given axis by a scalar or another curve.

fn multiply(&mut self, by_what: T)

Multiplies the Y values of the curve by a scalar or another curve.

fn add(&mut self, what: T)

Adds a scalar or another curve to the Y values of the curve.