zarza_etc::curve
Trait CurveOperations

pub trait CurveOperations {

    // Required methods
    fn is_oob(&self, x: NotNan<f64>) -> bool;
    fn bounds(&self) -> Option<(NotNan<f64>, NotNan<f64>)>;
    fn set_units(&mut self, axis: CurveAxis, units: String);
    fn integral(&self) -> f64;
    fn dist_mean(&self) -> f64;
    fn integrate(&mut self, k: f64);
    fn flip(&mut self);
    fn extend_left(&mut self) -> Result<(), String>;
    fn extend_right(&mut self) -> Result<(), String>;
    fn invert_axis(&mut self, axis: CurveAxis, factor: NotNan<f64>);
    fn x_points(&self) -> Vec<NotNan<f64>>;
    fn set(&mut self, x: NotNan<f64>, y: f64);
    fn get_point(&self, x: NotNan<f64>) -> f64;
    fn get_diff(&self, x: NotNan<f64>) -> f64;
    fn assign(&mut self, other: &Curve);
    fn from_existing(&mut self, other: &Curve, y_units: f64);
    fn clear(&mut self);
    fn debug(&self);
    fn load_curve(&mut self, path: &str) -> Result<(), Box<dyn Error>>;
    fn load_slice_dat(
        &mut self,
        path: &str,
        row_index: usize,
    ) -> Result<(), Box<dyn Error>>;
    fn save_curve(&self, path: &str) -> Result<(), Box<dyn Error>>;

}

Trait defining operations that can be performed on a Curve.

Required Methods

fn is_oob(&self, x: NotNan<f64>) -> bool

Check if a given x value is out of bounds.

fn bounds(&self) -> Option<(NotNan<f64>, NotNan<f64>)>

Returns the minimum and maximum X values, if any.

fn set_units(&mut self, axis: CurveAxis, units: String)

Set the units for a given axis.

fn integral(&self) -> f64

Calculate the integral of the curve.

fn dist_mean(&self) -> f64

Computes the mean value of the curve’s distribution.

fn integrate(&mut self, k: f64)

Integrates the curve, starting from a given value f, and modifying it in place.

fn flip(&mut self)

Flips the curve, swapping X and Y axes and their units.

fn extend_left(&mut self) -> Result<(), String>

Extends the curve to the left using the first defined point.

fn extend_right(&mut self) -> Result<(), String>

Extends the curve to the right using the last defined point.

fn invert_axis(&mut self, axis: CurveAxis, factor: NotNan<f64>)

Inverts the values of a given axis, multiplying by a factor.

fn x_points(&self) -> Vec<NotNan<f64>>

Returns a vector of all X points in the curve.

fn set(&mut self, x: NotNan<f64>, y: f64)

Sets a point in the curve, adding it if it doesn’t exist.

fn get_point(&self, x: NotNan<f64>) -> f64

Gets the Y value for a given X, interpolating if necessary.

fn get_diff(&self, x: NotNan<f64>) -> f64

Gets the derivative (slope) of the curve at a given X.

fn assign(&mut self, other: &Curve)

Assigns values from another curve, interpolating as necessary.

fn from_existing(&mut self, other: &Curve, y_units: f64)

Copies values from another curve, scaling Y values by a factor.

fn clear(&mut self)

Clears the curve, removing all points and resetting OOB values.

fn debug(&self)

Prints debug information about the curve.

fn load_curve(&mut self, path: &str) -> Result<(), Box<dyn Error>>

Loads curve data from a CSV file.

fn load_slice_dat( &mut self, path: &str, row_index: usize, ) -> Result<(), Box<dyn Error>>

Loads curve data from a specific row in a CSV file with headers. Only works for CSV files with slice format.

fn save_curve(&self, path: &str) -> Result<(), Box<dyn Error>>

Saves the curve data to a CSV file.