zarza_etc::detector
Struct Detector

pub struct Detector { /* private fields */ }

Main Detector structure

Implementations

impl Detector

Implementations of methods for Detector.

pub fn new() -> Detector

Create a new Detector instance with default values and load properties from a file.

pub fn properties(&self) -> &DetectorProperties

Inmutable access to the detector properties.

pub fn get_spec(&self) -> DetectorSpec

Get a clone of the current detector specification.

pub fn set_pixel_photon_flux(&mut self, flux_obj: Spectrum, flux_sky: Spectrum)

Set the photon flux per pixel for both the object and sky, and recalculate dependent values.

pub fn set_exposure_time(&mut self, t: f64)

Set the exposure time and recalculate dependent values.

pub fn set_detector(&mut self, det: String) -> bool

Set the detector by name from the properties. Returns true if successful, false if the detector name is not found.

pub fn dark_electrons(&self, t: f64) -> f64

Calculate the number of dark electrons based on the detector temperature and exposure time.

pub fn recalculate(&mut self)

Recalculate all dependent values based on the current detector specifications, exposure time, and photon fluxes.

pub fn signal_value(&self, wl: NotNan<f64>) -> f64

Get the signal value at a specific wavelength.

pub fn signal_crv(&self) -> Spectrum

Get a clone of the signal curve.

pub fn electrons_value(&self, wl: NotNan<f64>) -> f64

Get the electrons value at a specific wavelength.

pub fn electrons_crv(&self) -> Spectrum

Get a clone of the electrons curve.

pub fn read_out_noise(&self) -> f64

Get the read out noise in ADU.

pub fn noise_value(&self, wl: NotNan<f64>) -> f64

Get the noise value at a specific wavelength.

pub fn noise_crv(&self) -> Spectrum

Get a clone of the noise curve.

pub fn snr_value(&self, wl: NotNan<f64>) -> f64

Get the signal-to-noise ratio value at a specific wavelength.

pub fn snr_crv(&mut self) -> &mut Curve

Get a mutable reference to the signal-to-noise ratio curve.

pub fn get_properties(&self) -> &DetectorProperties

pub fn get_detector(&self) -> &DetectorSpec

pub fn get_exposure_time(&self) -> &f64

pub fn get_obj_photon_flux_per_pixel(&self) -> &Spectrum

pub fn get_sky_photon_flux_per_pixel(&self) -> &Spectrum

pub fn get_obj_photons_per_pixel(&self) -> &Spectrum

pub fn get_sky_photons_per_pixel(&self) -> &Spectrum

pub fn get_obj_electrons_per_pixel(&self) -> &Spectrum

pub fn get_sky_electrons_per_pixel(&self) -> &Spectrum

pub fn get_electrons_per_pixel(&self) -> &Spectrum

pub fn get_signal(&self) -> &Spectrum