zarza_etc
Module instrument_model

Structs

InstrumentModel

Structure representing the instrument model, including properties and various curves for dispersion and resolution.

InstrumentProperties

Structure to hold instrument properties loaded from configuration files and used in the instrument model.

Enums

InstrumentArm

Enum to represent the two arms of the instrument.

Constants

CAHA_APERTURE_AREA

Aperture area of the CAHA telescope in square meters.

CAHA_APERTURE_DIAMETER

Diameter of the CAHA telescope aperture in meters.

CAHA_EFFECTIVE_AREA

Effective area of the CAHA telescope in square meters.

CAHA_FOCAL_LENGTH

Focal length of the CAHA telescope in meters.

SPECTRAL_PIXEL_LENGTH

Spectral pixel length (number of pixels in the spectral direction).

TARSIS_SLICES

Number of slices in TARSIS instrument.