zarza_etc::simulation
Struct Simulation

pub struct Simulation { /* private fields */ }

Structure to hold the entire simulation state.

Implementations

impl Simulation

Implementation of methods for the Simulation struct.

pub fn get_input(&self) -> &Spectrum

Inmutable access to the input spectrum.

pub fn get_input_mut(&mut self) -> &mut Spectrum

Mutable access to the input spectrum.

pub fn get_filter(&self) -> &Curve

Inmutable access to the filter curve.

pub fn get_filter_equiv_bw(&self) -> &f64

Inmutable access to the filter equivalent bandwidth.

pub fn get_sky(&self) -> &Spectrum

Inmutable access to the sky spectrum.

pub fn get_sky_model(&self) -> &SkyModel

Inmutable access to the sky model.

pub fn get_tarsis_model(&self) -> &InstrumentModel

Inmutable access to the TARSIS instrument model.

pub fn get_tarsis_model_mut(&mut self) -> &mut InstrumentModel

Mutable access to the TARSIS instrument model.

pub fn get_det(&self) -> &Detector

Inmutable access to the detector model.

pub fn get_det_mut(&mut self) -> &mut Detector

Mutable access to the detector model.

pub fn get_params(&self) -> &SimulationParams

Inmutable access to the simulation parameters.

pub fn get_params_mut(&mut self) -> &mut SimulationParams

Mutable access to the simulation parameters.

pub fn new() -> Simulation

Constructor for Simulation with default values.

pub fn set_filter(&mut self, filter_path: &String)

Set the filter curve from a file path.

pub fn set_input_user(&mut self, spec_file_path: &String)

Set the input spectrum from a user-provided file path.

pub fn set_input_template(&mut self, template_path: &String)

Set the input spectrum from a predefined template.

pub fn set_input_line( &mut self, central_wl: NotNan<f64>, int_flux: f64, fwhm: f64, cont_flux: f64, line_type: &str, res: &str, arm: InstrumentArm, wl_min: f64, wl_max: f64, )

Set the input spectrum as a Gaussian emission or absorption line.

pub fn set_spatial_distribution( &mut self, spatial_distr: SpatialDistribution, pixel_position: usize, )

Set the spatial distribution of the input source flux.

pub fn normalize_to_filter_mag(&mut self, mag: f64)

Normalize the input spectrum to a given magnitude in the selected filter band. It is necessary to previously use โset_filterโ.

pub fn set_params(&mut self, params: SimulationParams)

Set the simulation parameters and update the sky spectrum and detector configuration accordingly.

pub fn simulate_arm(&mut self, arm: InstrumentArm)

Run the simulation for the specified instrument arm (blue or red).

pub fn read_out_noise(&self) -> f64

Get the read-out noise of the detector.

pub fn gain(&self) -> f64

Get the gain of the detector.

pub fn px_to_wavelength(&mut self, arm: InstrumentArm, px: NotNan<f64>) -> f64

Convert pixel position to wavelength using the instrument model.

pub fn wl_to_pixel_curve(&mut self, arm: InstrumentArm) -> Curve

Convert wavelength to pixel position using the instrument model.

pub fn signal(self, wl: NotNan<f64>) -> f64

Get the signal value at a given wavelength.

pub fn noise(self, wl: NotNan<f64>) -> f64

Get the noise value at a given wavelength.

pub fn electrons(self, wl: NotNan<f64>) -> f64

Get the number of electrons at a given wavelength.

pub fn texp_from_snr(&mut self, snr: f64, lambda_ref: NotNan<f64>) -> f64

Returns DIT or NDIT depending in what the user providad (first element of the touple), as well as INT (=DIT*NDIT) (second element of the touple)

pub fn lim_mag(&mut self, lambda_ref: NotNan<f64>, arm: InstrumentArm) -> f64

Calculate the limiting magnitude for a given wavelength and instrument arm.

pub fn lim_flux(&mut self, lambda_ref: NotNan<f64>, arm: InstrumentArm) -> f64

Calculate the limiting flux for a given wavelength and instrument arm.

pub fn rad_vel_unc(&mut self, arm: InstrumentArm, obj_size: f64) -> f64

Calculate the radial velocity uncertainty for a given instrument arm and object size.