zarza_etc::spectrum
Struct Spectrum

pub struct Spectrum { /* private fields */ }

A generic spectrum structure built upon a Curve structure. The “spectrum” encapsulates a “Curve” and provides additional operations for spectra, such as scaling by a factor or another curve, and axis inversion consistent with physical units. Spectra are commonly used to represent distributions (e.g., intensity, energy) across a domain such as wavelength or frequency. This struct enables both general curve manipulation and domain-specific spectral operations.

Implementations

impl Spectrum

Accessor methods for the Spectrum struct.

pub fn get_curve_mut(&mut self) -> &mut Curve

Mutable access to the underlying curve.

pub fn get_curve(&self) -> Curve

Immutable access to the underlying curve.